requirements.txt
absl-py==2.0.0
addict==2.4.0
albumentations==1.3.1
basicsr==1.4.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.5
easydict==1.11
facexlib==0.3.0
filelock==3.13.1
filterpy==1.4.5
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.44.0
fsspec==2023.10.0
future==0.18.3
gfpgan==1.3.8
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
httplib2==0.22.0
humanfriendly==10.0
idna==3.4
imageio==2.32.0
importlib-metadata==6.8.0
insightface==0.7.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.41.1
lmdb==1.4.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.1
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.16.3
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.0.1
platformdirs==4.0.0
prettytable==3.9.0
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
PyYAML==6.0.1
qudida==0.0.4
realesrgan==0.3.0
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sympy==1.12
tb-nightly==2.16.0a20231120
tensorboard-data-server==0.7.2
tf-keras-nightly==2.16.0.dev2023112010
threadpoolctl==3.2.0
tifffile==2023.9.26
tomli==2.0.1
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
wcwidth==0.2.9
Werkzeug==3.0.1
yapf==0.40.2
zipp==3.17.0
